@@ -22,20 +22,19 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "7.2.7",
+    "@angular/animations": "8.0.2",
     "@angular/cdk": "7.3.3",
-    "@angular/common": "7.2.7",
-    "@angular/compiler": "7.2.7",
-    "@angular/core": "7.2.7",
+    "@angular/common": "8.0.2",
+    "@angular/compiler": "8.0.2",
+    "@angular/core": "8.0.2",
     "@angular/flex-layout": "7.0.0-beta.22",
-    "@angular/forms": "7.2.7",
-    "@angular/http": "7.2.7",
+    "@angular/forms": "8.0.2",
     "@angular/material": "7.3.3",
     "@angular/material-moment-adapter": "7.3.3",
-    "@angular/platform-browser": "7.2.7",
-    "@angular/platform-browser-dynamic": "7.2.7",
-    "@angular/platform-server": "7.2.7",
-    "@angular/router": "7.2.7",
+    "@angular/platform-browser": "8.0.2",
+    "@angular/platform-browser-dynamic": "8.0.2",
+    "@angular/platform-server": "8.0.2",
+    "@angular/router": "8.0.2",
     "@ngrx/effects": "7.3.0",
     "@ngrx/entity": "7.3.0",
     "@ngrx/router-store": "7.3.0",
@@ -52,19 +51,19 @@
     "hammerjs": "2.0.8",
     "lodash": "4.17.11",
     "moment": "2.24.0",
-    "rxjs": "^6.4.0",
+    "rxjs": "^6.5.2",
     "ts-loader": "^5.3.1",
     "tslib": "^1.9.0",
-    "web-animations-js": "2.3.1",
-    "zone.js": "^0.8.29"
+    "web-animations-js": "github:angular/web-animations-js#release_pr208",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.11.0",
+    "@angular-devkit/build-angular": "~0.800.0",
     "@angular-devkit/core": "^7.1.0",
     "@angular-devkit/schematics": "^7.1.0",
-    "@angular/cli": "7.3.3",
-    "@angular/compiler-cli": "7.2.7",
-    "@angular/language-service": "7.2.7",
+    "@angular/cli": "8.0.4",
+    "@angular/compiler-cli": "8.0.2",
+    "@angular/language-service": "8.0.2",
     "@angularclass/hmr": "2.1.3",
     "@types/crypto-js": "3.1.43",
     "@types/express": "4.16.0",
@@ -72,12 +71,13 @@
     "@types/jasminewd2": "~2.0.3",
     "@types/lodash": "4.14.118",
     "@types/node": "~10.12.10",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "^5.0.1",
     "concurrently": "4.1.0",
     "http-server": "^0.11.1",
     "jasmine-core": "~3.3.0",
     "jasmine-marbles": "0.4.0",
     "jasmine-spec-reporter": "~4.2.1",
+    "json-server": "^0.15.0",
     "karma": "~3.1.1",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "^2.0.4",
@@ -88,7 +88,7 @@
     "rxjs-tslint-rules": "~4.13.0",
     "ts-node": "~7.0.1",
     "tslint": "~5.11.0",
-    "typescript": "3.2.4",
+    "typescript": "3.4.5",
     "udk": "^1.0.0",
     "webpack-cli": "^3.1.0",
     "webpack-node-externals": "1.7.2"
