@@ -50,7 +50,7 @@ export class ChipsListComponent implements ControlValueAccessor, OnInit {
   // fix the lint complaints about using a reference in template
   // It seems tslint requires now that a `@ViewChild` need to be declared
   // 'you are using blablabla that you're trying to access does not exist in the class declaration.'
-  @ViewChild('autoMember') autoMember: MatAutocomplete;
+  @ViewChild('autoMember', { static: true }) autoMember: MatAutocomplete;
   @Input() multiple = true;
   @Input() label = '添加/修改成员';
   @Input() placeholderText = '请输入成员 email';
